+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableShortArray.kt
@@ -233,6 +233,11 @@ public value class ImmutableShortArray @PublishedApi internal constructor(
             values.none(predicate)
 
     /**
+     * See [ShortArray.count]
+     */
+    public inline fun count(predicate: (predicate: Short) -> Boolean): Int = values.count(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableShortArray with all elements sorted
      * according to the natural sort order of the value returned by the [selector].
      *